Handle long integer conversion in NEC portinfo validation tests
There are a couple of tests for NEC portinfo validation that pass an id
which gets converted to a base-16 integer and then that gets converted
to a hex string. The problem is the integer conversion can make it a
long integer and then the hex built-in appends 'L' to the string, and
the tests do not account for that. This patch handles that case.

Closes-Bug: #1228212

Change-Id: Id3aed686329dae9cc4777213a4d2bcf406652e15