Remove _delete_port_security_group_bindings from delete_port
Since SecurityGroupPortBinding has a foreign key constraint with Port
and cascade deleting is defined, SecurityGroupPortBinding can be
deleted when port is deleted, thus _delete_port_security_group_bindings
call in ML2 delete_port is not needed and can be removed. Redundant
calls in other core plugins are also removed in this patch.

_delete_port_security_group_bindings is needed when updating ports so
this method cannot be removed directly.

Change-Id: I33c787fbed46d6049d3b24ccd246a8484d8aa306
Closes-Bug: #1373199