Add and use SecurityGroupAgentRpc
Add a new class, SecurityGroupAgentRpc, which is based on
SecurityGroupAgentRpcMixin.  Most uses of SecurityGroupAgentRpcMixin
follow the same pattern, so this class makes it possible to cut
down on some duplicated code.

Make use of SecurityGroupAgentRpc in: linuxbridge, openvswitch, mlnx,
nec, ofagent, oneconvergence, sriovnicagent, bigswitch, and hyperv.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I2aaa55e8e539e47427e56b4da42321cfcfcde622