Skip test_router_add_gateway_no_subnet for OpenFlow router
In IPv6 router, it is valid that router gateway port has no subnet.
It is introduced in the Neutron master in blueprint ipv6-router.
We need to skip the test until we have reasonable support of
IPv6 with LLA.

Related review is https://review.openstack.org/#/c/156283/
patch set 8 has some information on how to fix in NEC plugin
but it needs more investigation.

Change-Id: I09a061be5f57f6547bc37e37024610ec3fd16aca