Use 'if val' / 'if not val' for boolean logic
If there is no special reason to compare a variable with
True/False/None, python developers usually use:

* 'if val' rather than 'if val is True' or 'if val is not None'
* 'if not val' rather than 'if val is False' or 'if val is None'

Of course there is a case where we need to distinguish
True/False/None, for example, tri-state logic, or
a case where None is used to indicate it is a default.

This commit fixes other cases.

Change-Id: Id6f65584e285442708c2c2963b6d13af032e8eff