Remove unnecessary len(xxx) for list or dict
If python, bool([]) and bool({}) is False.
Thus when we need to check a list or a dict has at least
one member, we can do just 'if <list>' or 'if <dict>'.
There is no need to use len().

Similarly, 'if len(...) == 0' can be written as 'if not <list>'
or 'if not <dict>'.

Change-Id: Ifae79093667d68a4847b68cbe99e5c5e55b422a4