Enable pylint rules: too-many-{instance-attributes,public-methods}
The followig pylint rules are enabled.
* too-many-instance-attributes
* too-many-public-methods

[too-many-instance-attributes]
NECNWANeutronAgent has many instance attributes, but at a glance
it is not easy to clean up them. Let's mark this as an exception
of this rule.

[too-many-public-methods]
NwaClient defines public methods for all features in a single class.
It is not good as a design, but it cannot be addressed without
refactoring. As an exception, this rules is disbabled for NwaClient.

Change-Id: I75aaf9f9d1ba82617293dbfdaf348e32eb1713ae