Simplify agent/proxy_* logic using AgentProxyException
In the previous logic, most return values in agent/proxy_* consists
of a flag of success/failure and an actual return value.
As a result, the logic has many 'if <failure>' and return statement
and tends to be longer.

This commit introduces a new exception AgentProxyException and
the logic in agent/proxy_* catches this exception in failure cases.
To do this catch_exception_and_update_tenant_binding decorator was
introduced.

Change-Id: Ic452a28bee997aabe6357936d593ca5b368d1ef7