Cleanup TODO comments
Valid TODOs have been registered as launchpad bugs.
* test_proxy_tenant: bug 1548810
* test_proxy_l2: 1548144
* nwalib.client: 1547928

test_mech_necnwa: It is not a good idea to copy TODO from neutron
code because there is no chance to fix it. The code should be
managed by maintainers of the repo.

l2.utils: TODO exists but it seems okay to have it in utils.py
because l2.db_api module defines methods for nwa.l2.models.

Change-Id: If19c40a7959601e3003d95388af499d387244f57