Fixed bug to use old vlan id in multi-pod.
When vm1 is deleted, the nwa plugin requests to nwa with
delete_geneal_dev, and accepts the response of nwa. Next,
the nwa plugin delete the network segment information if
the port deleted is latest one.

When vm3 is created, the nwa plugin allocates a network segment
information as segmentation_id is zero. And segmentaion_id is set
after create_general_dev responsed.

In case of the bug, the segment information is same area vm1 and
vm2. Then vm3 is creating, it equals that segmentation_id is zero,
the segment information shouldn't be deleted when it accepts the
response of vm1's delete_general_dev.

Change-Id: I2fe86c54063310497ca0091200433e54bb16b198
Closes-Bug: #1569323