Fixed bug that old VLAN ID is used in multi-POD environment
When a network spreads multiple PODs and port creation is required
just after port deletion for a same POD, allocation of VLAN ID for
a network segment (POD) can be racy. As a result, a corresponding
port creation on NWA fails because VLAN segment is released during
the operation.

To fix the bug, this commit makes the following changes:
* NWA plugin now deletes VLAN ID information from neutron
  DB before calling delete request to NWA.
* When the delete request completes and if VLAN ID of the requested
  segment is zero (which means another allocation request is ongogin),
  VLAN ID information is not touched.

Note that this bug does not occurs in a single POD environemnt.

Change-Id: I2fe86c54063310497ca0091200433e54bb16b198
Closes-Bug: #1569323