Fixes limitation of nwa create tenant
There are two limitations.
The one is on RPC and the other is on DB field size.

The RPC limitation is that create_tenant_rpc_server did not return.
create_tenant_rpc_server calls server.start(), but it did not return
until the thread of callback is finished. Therefore, spawn an another
thread for handling callback.

The DB field size limitation is due to the size of information reported
in PluginReportStateAPI.  nwa agent reports a list of tenant IDs.  When
the number of active tenants becomes more than about 60, the size of
information exceeds the column size.  To reduce the size of information,
the number of tenants is now reported.

Change-Id: I332bff124a708eb41d272d95395bae8c7a028a69
Closes-Bug: #1569628