Change passing session to context in segments db functions
We have a number of functions that expect to get session as one
the arguments. Passing session is not correct and this prevents
of usage new enginefacade which expect context to be passed and
session will be injected in current context.

This change is based on change in Neutron, so here all affected
places are fixed to pass context.

Import of neutron exceptions in [1] change for import of exception
from neutron-lib to pass tests.

Also import segments_db instead of ml2_db to pass tests.

[1] - networking_nec/tests/unit/nwa/l3/test_plugin.py

Depends-On: I68ac442f11a6698bc43210fae877ba3809b1c02d

Partially-Implements blueprint: enginefacade-switch

Change-Id: Ia4e124898ae61fe6e4d23d93334d2f64edf4b389